Move form/types.ts to form/types/form-types.ts
Aligns with the table/types/ convention — a types directory per
feature, one file per concern. This creates room for config-types.ts
in a follow-up without turning form/types.ts into a dumping ground.
