+++ javascript/packages/core/components/form/components/form-dialog/types.ts
@@ -1,5 +1,5 @@
 import type { Size as DialogSize } from 'baseui/dialog';
-import type { FormData, FormProps } from '#core/components/form/types';
+import type { FormData, FormProps } from '#core/components/form/types/form-types';
 
 export interface FormDialogProps<FieldValues extends FormData = FormData>
   extends Pick<FormProps<FieldValues>, 'onSubmit' | 'initialValues' | 'children'> {
